Add arguments and formatting to messages
This also adds a few missing messages for certain errors, such as
listing other's keys and attempting to take a quantity of keys greater
than the user's current balance.

Also updated to api 12

+++ src/main/java/dev/flashlabs/cratecrate/command/location/Delete.java
@@ -1,7 +1,10 @@
 package dev.flashlabs.cratecrate.command.location;
 
+import dev.flashlabs.cratecrate.CrateCrate;
 import dev.flashlabs.cratecrate.command.CommandUtils;
+import dev.flashlabs.cratecrate.component.Crate;
 import dev.flashlabs.cratecrate.internal.Storage;
+import net.kyori.adventure.audience.Audience;
 import net.kyori.adventure.identity.Identity;
 import net.kyori.adventure.text.Component;
 import org.spongepowered.api.command.Command;
@@ -9,9 +12,11 @@ import org.spongepowered.api.command.CommandResult;
 import org.spongepowered.api.command.exception.CommandException;
 import org.spongepowered.api.command.parameter.CommandContext;
 import org.spongepowered.api.command.parameter.Parameter;
+import org.spongepowered.api.util.locale.LocaleSource;
 import org.spongepowered.api.world.server.ServerLocation;
 
 import java.sql.SQLException;
+import java.util.Optional;
 
 public final class Delete {
     public static final Component USAGE = CommandUtils.usage(
@@ -30,13 +35,28 @@ public final class Delete {
     private static CommandResult execute(CommandContext context) throws CommandException {
         var location = context.requireOne(Parameter.key("location", ServerLocation.class));
         location = location.withBlockPosition(location.blockPosition());
+
+        if (!Storage.LOCATIONS.containsKey(location)) {
+            throw new CommandException(CrateCrate.get().getMessage("command.location.delete.invalid-location", ((LocaleSource) context.cause().audience()).locale(),
+                    "location", location.worldKey().toString() + " " + location.position()
+            ));
+        }
+        //TODO: Get registered crate id from database?
+        Optional<Crate> crate = Storage.LOCATIONS.get(location);
         try {
             Storage.deleteLocation(location);
-            Storage.LOCATIONS.remove(location);
-            context.sendMessage(Identity.nil(), Component.text("Successfully deleted location."));
+            CrateCrate.get().sendMessage((Audience & LocaleSource) context.cause().audience(), "command.location.delete.success",
+                    "location", location.worldKey().asString() + " " + location.position(),
+                    "crate", crate.map(a -> a.id()).orElse("unavailable")
+            );
         } catch (SQLException e) {
-            throw new CommandException(Component.text("Error deleting location: " + e.getMessage()));
+            e.printStackTrace();
+            throw new CommandException(CrateCrate.get().getMessage("command.location.delete.failure", ((LocaleSource) context.cause().audience()).locale(),
+                    "location", location.worldKey().asString() + " " + location.position(),
+                    "crate", crate.map(a -> a.id()).orElse("unavailable")
+            ));
         }
+
         return CommandResult.success();
     }
